requirements.txt
asgiref==3.2.10
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
Django==3.1.1
django-cors-headers==3.5.0
django-crispy-forms==1.7.2
djangorestframework==3.11.1
get==2019.4.13
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
lxml==4.4.2
mysql-connector-python==8.0.20
numpy==1.18.1
pandas==1.0.1
parsel==1.5.2
Pillow==7.0.0
post==2019.4.13
Protego==0.1.16
protobuf==3.12.2
psycopg2==2.8.6
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyodbc==4.0.30
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.1
query-string==2019.4.13
queuelib==1.5.0
request==2019.4.13
requests==2.23.0
Scrapy==2.1.0
scrapy-user-agents==0.1.1
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
Twisted==20.3.0
ua-parser==0.10.0
urllib3==1.25.8
user-agents==2.1
w3lib==1.21.0
zope.interface==5.1.0
